Apply query parameter to ticket page. Move ticketdata endpoint to TicketController and apply authentication to it.
